requirements.txt
watchdog>=2.1.9
PyYAML>=6.0
rich>=13.0.0
click>=8.0.0
pandas>=1.5.0
matplotlib>=3.5.0
plotly>=5.0.0
psutil>=5.9.0
